+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Vector distance squared static method.
 - Circle from vector static constructor method.
 
+### Changed
+
+- Refactor object instantiation in Circle tests.
+
 ## [5.2.0] - 2025-01-20
 
 ### Added